requirements.txt
pymongo[srv]==4.15.4
openai==2.0.1
requests==2.32.5
beautifulsoup4==4.14.2
Flask==3.0.3
flask-cors==4.0.1
gunicorn==23.0.0
sentence-transformers>=3.0.0
scikit-learn==1.3.2
Flask-Limiter==3.11.0
redis==7.0.1
python-dotenv==1.0.0
chromadb>=0.4.0
# Neo4j Python Driver - required for knowledge_graph connection
neo4j>=5.0.0